+++ prisma/migrations/20210207093031_init/migration.sql
@@ -1,0 +1,13 @@
+-- CreateTable
+CREATE TABLE "User" (
+    "id" INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT,
+    "lineUserId" TEXT NOT NULL,
+    "displayName" TEXT NOT NULL,
+    "profileUrl" TEXT NOT NULL,
+    "statusMessage" TEXT NOT NULL,
+    "createdAt" DATETIME NOT NULL,
+    "updatedAt" DATETIME NOT NULL
+);
+
+-- CreateIndex
+CREATE UNIQUE INDEX "User.lineUserId_unique" ON "User"("lineUserId");