+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# 0.3.0 (2021-02-17)
+
+
+### Features
+
+* add AllExceptionFilter, replacing PrismaFilter ([40733d0](https://github.com/Howard86/line-blockchain/commit/40733d046336f323384cec2dd0dec0790f0c10bf))
+* add Database module ([4bae2b9](https://github.com/Howard86/line-blockchain/commit/4bae2b95be6af3b52eb68d140dfb12ede21330c9))
+* add global PrismaFilter ([61538da](https://github.com/Howard86/line-blockchain/commit/61538da2049b7bf24bcf62e59ec66da86f89f881))
+* add NullCheckInterceptor for findOne service ([7de969f](https://github.com/Howard86/line-blockchain/commit/7de969f774e54e3a7063673bbd9e5191b4853118))
+* set up ConfigModule with .env ([5084cec](https://github.com/Howard86/line-blockchain/commit/5084cec3e4b7aefa5502d860081b87e8831e9256))
+* set up LIFF starter app with hbs ([4a85311](https://github.com/Howard86/line-blockchain/commit/4a85311cd5e2e54b62436241500a319caf949ab9))
+* set up prisma schema & migration and npm scripts ([1164684](https://github.com/Howard86/line-blockchain/commit/1164684ca8e1dd5f5648b5c2827cae1f09b71808))
+* set up validation pipe on DTOs ([4cf19fb](https://github.com/Howard86/line-blockchain/commit/4cf19fbb47262942ef16976174b310ed414f4376))
+* update users controller with new users service ([9ea5921](https://github.com/Howard86/line-blockchain/commit/9ea59212188a0a7cb73885942a2f310e15a361a9))
+* update users service with database module ([eb957f8](https://github.com/Howard86/line-blockchain/commit/eb957f8f971ddb86c18480f2dbb93fdd98359c41))
+
+
+
 # [0.2.0](https://github.com/Howard86/line-blockchain/compare/v0.1.0...v0.2.0) (2021-02-06)